+++ R/computational_grid.R
@@ -1,9 +1,26 @@
-create_computation_grid <- function(data, hparams, formulas, fold_cols, n_hparam_combinations = NULL){
+create_computation_grid <- function(data, hparams, formulas, fold_cols){
+
+  # Hyperparameters
+
+  # If no hparams were passed,
+  # create NA list
+  if (is.null(hparams)){
+    hparams <- list(".__NA__" = NA)
+  }
+
+  # Extract n_hparam_combinations
+  if (".n" %in% names(hparams)){
+    n_hparam_combinations <- hparams[[".n"]]
+    hparams[[".n"]] <- NULL
+  } else {
+    n_hparam_combinations <- NULL
+  }
 
   hparams_grid <- create_hparams_grid(hparams, n = n_hparam_combinations)
   hparam_combination <- hparams_grid[["hparam_combination"]]
 
-  # Not made yet, just temporary output
+
+  # Create folds map
   folds_grid <- create_folds_grid(data = data, fold_cols = fold_cols)
 
   # Just the enframed formulas vector
@@ -31,10 +48,12 @@ create_hparams_grid <- function(hparams, n = NULL){
   # hparam combinations, if specified
   if (!is.null(n)){
 
-    if (is_between_(n)){
-      sampler_fn <- dlyr::sample_frac
+    if (is_between_(n, 0, 1)){
+      sampler_fn <- dplyr::sample_frac
     } else {
-      sampler_fn <- dlyr::sample_n
+      if (n > nrow(grid))
+        stop("'n' was greater than the rows in the hparams grid.")
+      sampler_fn <- dplyr::sample_n
     }
 
     grid <- grid %>%
@@ -47,17 +66,7 @@ create_hparams_grid <- function(hparams, n = NULL){
 }
 
 create_folds_grid <- function(data, fold_cols){
-
   # Create folds map
   folds_map_and_n_folds <- create_folds_map(data, fold_cols)
   folds_map_and_n_folds[["folds_map"]]
 }
-
-
-
-# create_computation_grid(hparams = list("a" = c(-1,0,1),
-#                                        "b" = 7,
-#                                        "c" = c(5,2,4,5)),
-#                         formulas = c("y ~ x", "y ~ z"),
-#                         folds_map = NULL, # TODO use actual folds map
-#                         n_hparam_combinations = NULL)