+++ R/cross_validate_fn.R
@@ -38,14 +38,19 @@
 #'
 #'  Must have the following function arguments:
 #'
-#'  \code{function(train_data, formula)}
-#' @param predict_type The \code{type} argument for \code{\link[stats:predict]{predict()}}.
+#'  \code{function(train_data, formula, hyperparameters)}
+#' @param predict_fn Function for predicting the targets in the test folds using the fitted model object.
+#'  Will usually wrap \code{\link[stats:predict]{stats::predict()}}, but doesn't have to.
 #'
-#'  When the defaults fail, provide it such that the \code{\link[stats:predict]{predict()}}
-#'  output is as follows:
+#'  Must have the following function arguments:
+#'
+#'  \code{function(test_data, model, formula)}
+#'
+#'  Must return predictions in the following formats, depending on \code{type}:
 #'
 #'  \subsection{Binomial}{
-#'  Vector or one-column matrix / data frame with probabilities (0-1).
+#'  Vector or one-column matrix / data frame with probabilities (0-1)
+#'  of the second class, alphabetically.
 #'  E.g.:
 #'
 #'  \code{c(0.3, 0.5, 0.1, 0.5)}
@@ -70,19 +75,50 @@
 #'   0.368 \tab 0.322 \tab 0.310\cr
 #'   0.375 \tab 0.371 \tab 0.254\cr
 #'   ... \tab ... \tab ...}
-#'
-#'  N.B. \code{predict_fn} and \code{predict_type} are mutually exclusive. Specify only one of them.
 #'  }
-#' @param predict_fn Function for predicting the targets in the test folds using the fitted model object.
-#'  Will usually wrap \code{\link[stats:predict]{predict()}}, but doesn't have to.
-#'  Must return predictions in the format described in \code{predict_type} above.
+#' @param preprocess_fn Function for preprocessing the training and test sets.
+#'
+#'  Can, for instance, be used to standardize both the training and test sets with the scaling and centering parameters
+#'  from the training set.
 #'
 #'  Must have the following function arguments:
 #'
-#'  \code{function(test_data, model, formula = NULL)}
+#'  \code{function(train_data, test_data,}
 #'
-#'  N.B. \code{predict_fn} and \code{predict_type} are mutually exclusive. Specify only one of them.
+#'  \verb{         }\code{formula, hyperparameters)}
 #'
+#'  Must return a list with the preprocessed \code{train_data} and the \code{test_data}:
+#'
+#'  \code{list("train" = train_data,}
+#'
+#'  \verb{     }\code{"test" = test_data)}
+#'
+#'  N.B. When \code{preprocess_once} is FALSE, the current formula and
+#'  hyperparameters will be provided. Otherwise,
+#'  these arguments will be \code{NULL}.
+#' @param preprocess_once Whether to apply the preprocessing once
+#'  (ignoring the formula and hyperparameters arguments in \code{preprocess_fn})
+#'  or for every model separately. (Logical)
+#'
+#'  When preprocessing does not depend on the current formula or hyperparameters,
+#'  we can do the preprocessing of each train/test split once, to save time.
+#'  This \strong{may require holding a lot more data in memory} though, why
+#'  why it is not the default setting.
+#' @param hyperparameters List of hyperparameter values to cross-validate.
+#' Every combination will be created.
+#'
+#'  Add \code{".n"} to sample the combinations. Can be the number of combinations to use,
+#'  or a percentage between \code{0} and \code{1}.
+#'
+#'  E.g.
+#'
+#'  \code{list(".n" = 10,  # sample 10 combinations}
+#'
+#'  \verb{     }\code{"lrn_rate" = c(0.1, 0.01, 0.001),}
+#'
+#'  \verb{     }\code{"h_layers" = c(10, 100, 1000),}
+#'
+#'  \verb{     }\code{"drop_out" = runif(5, 0.3, 0.7))}
 #' @details
 #'
 #'  Packages used:
@@ -387,13 +423,17 @@
 #'                                fold_cols = ".folds",
 #'                                parallel = FALSE)})
 #' }
-cross_validate_fn <- function(data, model_fn, formulas,
+cross_validate_fn <- function(data,
+                              formulas,
+                              model_fn,
+                              predict_fn,
+                              preprocess_fn = NULL,
+                              preprocess_once = FALSE,
+                              hyperparameters = NULL,
                               fold_cols = '.folds',
                               type = 'gaussian',
                               cutoff = 0.5,
                               positive = 2,
-                              predict_type = NULL,
-                              predict_fn = NULL,
                               metrics = list(),
                               rm_nc = FALSE,
                               parallel = FALSE){
@@ -401,12 +441,14 @@ cross_validate_fn <- function(data, model_fn, formulas,
   return(custom_cross_validate_list(data = data,
                                     formulas = formulas,
                                     model_fn = model_fn,
+                                    predict_fn = predict_fn,
+                                    preprocess_fn = preprocess_fn,
+                                    preprocess_once = preprocess_once,
+                                    hyperparameters = hyperparameters,
                                     fold_cols = fold_cols,
                                     family = type,
                                     cutoff = cutoff,
                                     positive = positive,
-                                    predict_type = predict_type,
-                                    predict_fn = predict_fn,
                                     metrics = metrics,
                                     rm_nc = rm_nc,
                                     model_verbose = FALSE,